requirements.txt
telethon
anthropic
ccxt
python-telegram-bot>=20.0
python-dotenv
Pillow
aiohttp
aiofiles
